Move site field constants to their own file
